Refactor forms extracting common logic to parent
