+++ cosmetics-web/app/forms/sign_in_form.rb
@@ -1,6 +1,4 @@
-class SignInForm
-  include ActiveModel::Model
-  include ActiveModel::Attributes
+class SignInForm < Form
   include EmailFormValidation
 
   attribute :password
